refactor(codex): derive plugin config contracts from schemas
