+++ extensions/codex/src/app-server/config-parsing.ts
@@ -5,13 +5,12 @@ import { z } from "zod";
 import {
   CODEX_PLUGIN_MARKETPLACE_NAME_PATTERN,
   type CodexAppServerCommandSource,
-  type CodexPluginConfig,
   type CodexPluginDestructiveApprovalMode,
   type CodexPluginDestructivePolicy,
   type CodexPluginMarketplaceName,
   type ResolvedCodexPluginPolicy,
   type ResolvedCodexPluginsPolicy,
-} from "./config-contracts.js";
+} from "./config-contracts.shared.js";
 import { normalizeCodexServiceTier } from "./config-utils.js";
 import {
   codexDiscoveryConfigSchema,
@@ -189,7 +188,15 @@ const codexPluginConfigSchema = z
   })
   .strict();
 
-export function readCodexPluginConfig(value: unknown): CodexPluginConfig {
+export type ParsedCodexSupervisionEndpoint = z.infer<typeof codexSupervisionEndpointSchema>;
+export type ParsedCodexPluginConfig = Omit<
+  z.infer<typeof codexPluginConfigSchema>,
+  "codexPlugins"
+> & {
+  codexPlugins?: z.infer<typeof codexPluginsConfigSchema>;
+};
+
+export function readCodexPluginConfig(value: unknown): ParsedCodexPluginConfig {
   const appServer = asNullableRecord(asNullableRecord(value)?.appServer);
   if (appServer?.approvalPolicy === "untrusted") {
     throw new Error(
